Rename old node meta package
